+++ app/routes/components/CodeHighlight/index.tsx
@@ -1,9 +1,6 @@
-import { ComponentTemplate } from "@/components/ComponentTemplate";
-import { Section } from "@/components/Section";
-import { Typography } from "@/components/Typography";
-
+import { useState } from "react";
+import { DocLayout } from "@/layouts/DocLayout/doc-layout";
 import type { Route } from "./+types/index";
-
 import { changelog } from "./changelog.codegen";
 import Demo01 from "./demos/Demo01";
 import demo01SourceCode from "./demos/Demo01.source_code.codegen";
@@ -20,17 +17,38 @@ export function meta(_: Route.MetaArgs) {
 }
 
 export default function CodeHighlightPage() {
+  const [tabsValue, setTabsValue] = useState("doc");
+
+  const tabs = [
+    { name: "doc", label: "Document" },
+    { name: "source", label: "Source Code" },
+    { name: "changelog", label: "Changelog" },
+  ];
+
   return (
-    <ComponentTemplate component="CodeHighlight">
-      <Section.Root title="CodeHighlight">
-        <Section.Demo node={<Demo01 />} code={demo01SourceCode} />
-        <Typography.H2>See Also</Typography.H2>
-        <Typography.Ul>
-          <Typography.Li>CodeBlock.</Typography.Li>
-        </Typography.Ul>
-        <Section.SourceCode files={sourceCode} />
-        <Section.Changelog changelog={changelog} />
-      </Section.Root>
-    </ComponentTemplate>
+    <DocLayout>
+      <DocLayout.Title
+        title="CodeHighlight"
+        desc="Render your code."
+        tabsValue={tabsValue}
+        tabs={tabs}
+        onTabsValueChange={setTabsValue}
+      />
+      {tabsValue === "doc" && (
+        <DocLayout.Content>
+          <DocLayout.Live node={<Demo01 />} code={demo01SourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "source" && (
+        <DocLayout.Content>
+          <DocLayout.Files files={sourceCode} />
+        </DocLayout.Content>
+      )}
+      {tabsValue === "changelog" && (
+        <DocLayout.Content>
+          <DocLayout.Changelog changelog={changelog} />
+        </DocLayout.Content>
+      )}
+    </DocLayout>
   );
 }